instruction revisions
