+++ README.md
@@ -1,0 +1,23 @@
+# Boolean Operators Lab
+
+## Pre-Requisites
+- You are expected to review the `Working with Basic Operators` written lecture prior to working on this lab. If you are not familiar with the concepts being explored in the lab, please make sure to review the study material first.
+
+## Description
+In this exercise, you will practice using logical, or boolean, operators. Please attempt this execise without using conditional statements, such as if/else statements.
+
+## Completing the Lab
+You will be editing the file `Booleans.java`. This file can be found under the `src/main/java` package. Review the file's code and code comments for further instructions.
+
+Tasks to complete within the `Booleans.java` file:
+- [ ] Revise the return statement of the method named `returnBoolean` so that it uses a boolean operator to return the correct result.
+- [ ] Revise the return statement of the method named `returnNot` so that it uses a boolean operator to return the correct result.
+- [ ] Revise the return statement of the method named `returnAnd` so that it uses a boolean operator to return the correct result.
+- [ ] Revise the return statement of the method named `returnOr` so that it uses a boolean operator to return the correct result.
+
+## Running Tests
+If you look within the `src/test/java` folder, you will find the Java class `BooleansTest.java`. The `src/test/java` folder will contain your test files.  
+
+To run the tests, click on the Testing tab (test beaker icon on the left-hand side) and run the test cases by using the "Run Test" button (play icon). Please ensure that you do not already have the program running.
+
+You can optionally run the program for manual testing by switching to the file named `Main.java` under the `src/main/java` package and clicking the "Run Java" button (play icon) on the top-right of VSCode.
