+++ src/main/java/Booleans.java
@@ -1,0 +1,46 @@
+public class Booleans {
+
+    /**
+     * This method returns either true or false, depending on the parameter `b`.
+     * 
+     * @param b - a boolean value.
+     * @return - if b is true, return true. if b is false, return false.
+     */
+    public boolean returnBoolean(boolean b){
+        return false;
+    }
+
+    /**
+     * This method returns the opposite boolean value of the parameter `b`.
+     * 
+     * @param b - a boolean value.
+     * @return - if b is true, return false. if b is false, return true.
+     */
+    public boolean returnNot(boolean b){
+        return false;
+    }
+
+    /**
+     * This method returns either true or false, depending on whether or not the parameters both contain the value `true`.
+     * 
+     * @param b1 - a true/false value.
+     * @param b2 - a true/false value.
+     * @return - true if both b1 and b2 are true. false otherwise.
+     */
+    public boolean returnAnd(boolean b1, boolean b2){
+        return false;
+    }
+
+    /**
+     * This method returns either true or false, depending on whether or not either of the parameters contain the value `true` or not.
+     * 
+     * @param b1 - a true/false value.
+     * @param b2 - a true/false value.
+     * @return - true if either b1 or b2 are true. false if both are false.
+     */
+    public boolean returnOr(boolean b1, boolean b2){
+        return false;
+    }
+
+
+}
